# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QLGD API",
        "description": "Tài liệu OpenAPI cho hệ thống Quản lý Giảng dạy.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "Máy chủ API"
        }
    ],
    "paths": {
...
        "/api/lecturer/sessions/{id}/start": {
            "post": {
                "tags": [
                    "Lecturer - Buổi dạy"
                ],
                "summary": "Bắt đầu buổi dạy",
                "operationId": "lecturerSessionStart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 345
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bắt đầu buổi dạy thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 345
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "TEACHING"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền truy cập",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy buổi dạy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Trạng thái không hợp lệ để bắt đầu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lecturer/sessions/{id}/finish": {
            "post": {
                "tags": [
                    "Lecturer - Buổi dạy"
                ],
                "summary": "Kết thúc buổi dạy",
                "operationId": "lecturerSessionFinish",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 345
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hoàn tất buổi dạy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 345
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "DONE"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền truy cập",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy buổi dạy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Chưa điểm danh hoặc trạng thái không cho phép hoàn tất",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/training_department/approvals/leave/{leave}": {
            "post": {
                "tags": [
                    "Training Department - Approvals"
                ],
                "summary": "Phê duyệt / từ chối đơn nghỉ dạy",
                "operationId": "trainingApproveLeave",
                "parameters": [
                    {
                        "name": "leave",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "APPROVED hoặc REJECTED",
                                        "type": "string",
                                        "example": "APPROVED"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Đã sắp xếp giảng viên thay thế",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật trạng thái đơn nghỉ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LeaveRequestResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Thông tin xác thực không hợp lệ."
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ScheduleItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 345
                    },
                    "session_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-21"
                    },
                    "status": {
                        "type": "string",
                        "example": "PLANNED"
                    },
                    "start_time": {
                        "type": "string",
                        "example": "07:00"
                    },
                    "end_time": {
                        "type": "string",
                        "example": "08:45"
                    },
                    "assignment": {
                        "properties": {
                            "semester_label": {
                                "type": "string",
                                "example": "2025-2026 HK1"
                            },
                            "subject": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "CT101"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Cấu trúc dữ liệu"
                                    }
                                },
                                "type": "object"
                            },
                            "classUnit": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 88
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "LHP01"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Lớp HP 01"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "room": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 12
                            },
                            "code": {
                                "type": "string",
                                "example": "B3-201"
                            },
                            "name": {
                                "type": "string",
                                "example": "Phòng B3-201"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "LeaveRequestResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 100
                    },
                    "schedule_id": {
                        "type": "integer",
                        "example": 345
                    },
                    "lecturer_id": {
                        "type": "integer",
                        "example": 7
                    },
                    "reason": {
                        "type": "string",
                        "example": "Ốm đột xuất"
                    },
                    "status": {
                        "type": "string",
                        "example": "PENDING"
                    },
                    "approved_by": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "approved_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    },
                    "schedule": {
                        "$ref": "#/components/schemas/ScheduleItem"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter the access token returned after login (omit the Bearer prefix)."
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}